requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-ninja==1.5.0
GDAL==3.2.3
geopandas==1.1.1
idna==3.11
numpy==1.25.2
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic_core==2.41.5
pyogrio==0.11.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
